use size_t instead of int for array size

+++ src/t8_schemes/t8_default/t8_default_pyramid/t8_default_pyramid_cxx.cxx
@@ -480,7 +480,7 @@ t8_default_scheme_pyramid_c::t8_element_reference_coords (const t8_element_t
                                                           *elem,
                                                           const double
                                                           *ref_coords,
-                                                          const int
+                                                          const size_t
                                                           num_coords,
                                                           double *out_coords)
   const
